+++ lib/components/steps/StepTwo.tsx
@@ -1,87 +1,17 @@
-import { Box, Button, Checkbox, Chip, Divider, Grid, IconButton, Link, Tab, Tabs, Tooltip, Typography } from '@mui/material';
-import { ModalLayout } from "../layout/ModalLayout.tsx";
+import { Box, Button, Checkbox, Chip, Grid, IconButton, Link, Typography } from '@mui/material';
 import { useCdeContext } from "../../CdeContext.tsx";
 import React, { useState } from 'react';
 import { ArrowDropDown, ArrowIcon, CheckboxDefault, CheckboxSelected, ChevronRight, GlobeIcon, LeftIcon, LinkIcon, RightIcon } from '../../icons/index.tsx';
 
-function a11yProps(index: number) {
-    return {
-        id: `simple-tab-${index}`,
-        'aria-controls': `simple-tabpanel-${index}`,
-    };
-}
-
-const tabsArr = [
-    {
-        label: 'Select default repository',
-        heading: 'Select default repository',
-        description: 'Choose a default repository that best fit the dataset you have selected.'
-    },
-    {
-        label: 'Suggestions',
-        heading: 'Accept or decline suggestions',
-        description: 'Suggestions are based on what was previously mapped before.'
-    },
-    {
-        label: 'Map the rest of the dataset',
-        heading: 'Configure all selected dataset',
-        description: 'All (both mapped and unmapped) your selected dataset will be here to configure.'
-    }
-];
-
-function StepOne() {
+function StepTwo() {
     const { mapping } = useCdeContext();
 
-    const [value, setValue] = React.useState(1);
     const [showOtherSuggestions, setShowOtherSuggestions] = useState(false);
-
-    const handleChange = (_event: React.SyntheticEvent, newValue: number) => {
-        setValue(newValue);
-    };
 
     console.log(mapping)
 
     return (
-        <ModalLayout>
-            <Box sx={{
-                borderBottom: '0.0625rem solid #ECEDEE',
-                padding: '0 1.5rem',
-            }} display='flex' justifyContent='space-between' alignItems='center'>
-                <Tabs value={value} onChange={handleChange} aria-label="basic tabs example">
-                    {/* icon={index !== 0 ? <ChevronRight /> : ''} */}
-                    {tabsArr?.map((tab, index) => (
-                        <Tooltip
-                            placement='bottom'
-                            title={
-                                <>
-                                    <Typography sx={{
-                                        fontSize: '0.75rem',
-                                        fontWeight: 600,
-                                        lineHeight: '142.857%',
-                                        marginBottom: '0.25rem',
-                                        color: '#fff',
-                                    }}>{tab.heading}</Typography>
-                                    <Typography sx={{
-                                        fontSize: '0.75rem',
-                                        fontWeight: 400,
-                                        lineHeight: '142.857%',
-                                        color: '#fff',
-                                    }}>{tab.description}</Typography>
-                                </>
-                            }
-                        >
-                            <Tab label={`${index + 1}. ${tab.label}`} {...a11yProps(index)} />
-                        </Tooltip>
-                    ))}
-                </Tabs>
-
-                <Box display='flex' gap='0.625rem' alignItems='center'>
-                    <Button variant='text'>
-                        Continue without suggestions
-                    </Button>
-                </Box>
-            </Box>
-
+        <>
             <Box overflow='auto' maxHeight='calc(100vh - (3.9375rem + 3.5625rem + 4.4375rem + 2rem + 2rem))' p='1.5rem' pb={6}>
                 <Box mb={3} borderBottom='0.0625rem solid #ECEDEE' py='0.6875rem' display='flex' alignItems='center' justifyContent='space-between'>
                     <Typography sx={{
@@ -644,8 +574,8 @@ function StepOne() {
                     <Button variant='contained'>Accept selected mapping</Button>
                 </Box>
             </Box>
-        </ModalLayout>
+        </>
     );
 }
 
-export default StepOne;
+export default StepTwo;
